🔥 Mass Delete 18 (#6)
* Dropping `DefaultHornClausebase` and `DefaultHornClausebaseIndexer`

The `DefaultHornClauseContext` included a reference to the
`DefaultHornClausebase`, but this reference would instantiate
a new `DefaultHornClausebase` only if the current instance
was null. By construction, this was impossible since the only
public constructor threw an error if a context was not provided.

* Drop `-dropPos`, `-save`, and remove unreachable code

* Drop multiple keywords from the `FileParser`:

define
noMode
constraints
determinate
numericFunctionAsPred
booleanFunctionAsPred
categoricalFunctionAsPred
structuredFunctionAsPred
anythingFunctionAsPred
listOfNumericFunctionAsPred
listOfBooleanFunctionAsPred
listOfCategoricalFunctionAsPred
listOfStructuredFunctionAsPred
listOfAnythingFunctionAsPred
threshold
filter
relevantClause
relevantFeature
relevantMode
supportingLiteral
costFinal
isa
hiddenPred
weight
wgt
output
out
dontComplainAboutMultipleTypes
useLeadingQuestionMarkVariables
importIfExists

* Dropping "developer" parameters from Utils

There were a large number of parameters in the Utils which were
set based on the return value of `getUserName`-type methods.
To my knowledge, no person matching any of those user names has
developed on this codebase in some time.

* Small removal before jumping into `FileParser`

* Dropping the `inlines_.fopcLibrary` files

These files contained a single line saying:
`useStdLogicVariables: true`, which should do nothing.

* Drop keywords from `FileParser`

* Dropping commented-out lines in the .fopcLibrary files

* Advice Refactor

% [AdviceProcessor]  Generated 0 clauses at relevance level STRONGLY_IRRELEVANT.
occurs at every iteration, so this commit drops everything that was not reached
on the branch of the `if`-statement that resulted in the line shown above.